1.8版本，重构服务层
