+++ app/Http/Controllers/Pay/PaypalPayController.php
@@ -110,7 +110,7 @@ class PaypalPayController extends PayController
             $result = $payment->execute($execute, $paypal);
             $payData = $result->toArray();
             if ($payData['payer']['status'] == "VERIFIED" && $payData['transactions'][0]['amount']['currency'] == "USD") {
-                $this->successOrder($oid, $paymentId, $cacheord['actual_price']);
+                $this->orderService->successOrder($oid, $paymentId, $cacheord['actual_price']);
                 return redirect(site_url().'searchOrderById?order_id='.$oid);
             }
         } catch(\Exception $e) {
@@ -153,7 +153,7 @@ class PaypalPayController extends PayController
             $result = $payment->execute($execute, $paypal);
             $payData = $result->toArray();
             if ($payData['payer']['status'] == "VERIFIED" && $payData['transactions'][0]['amount']['currency'] == "USD") {
-                $this->successOrder($oid, $paymentId, $cacheord['actual_price']);
+                $this->orderService->successOrder($oid, $paymentId, $cacheord['actual_price']);
             }
         } catch(\Exception $e) {
            Log::info('paypal异常：' . $e->getMessage());
